backend/requirements.txt
Flask==3.0.0
python-pptx==0.6.23
gunicorn==21.2.0
flask-cors
lxml